LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/smooth-scroll-nonstop-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: test.step_timeout is not a function. (In 'test.step_timeout(() => {
      reject(`No ${eventName} event received for target ${target}`);
    }, timeoutMs)', 'test.step_timeout' is undefined...

Harness Error (FAIL), message = Unhandled rejection: test.step_timeout is not a function. (In 'test.step_timeout(() => {
      reject(`No ${eventName} event received for target ${target}`);
    }, timeoutMs)', 'test.step_timeout' is undefined)

PASS noop scrollTo doesn't interrupt ongoing smooth scroll.
TIMEOUT noop scrollIntoView doesn't interrupt ongoing smooth scroll. Test timed out

Harness Error (FAIL), message = Unhandled rejection: test.step_timeout is not a function. (In 'test.step_timeout(() => {
      reject(`No ${eventName} event received for target ${target}`);
    }, timeoutMs)', 'test.step_timeout' is undefined)

PASS noop scrollTo doesn't interrupt ongoing smooth scroll.
TIMEOUT noop scrollIntoView doesn't interrupt ongoing smooth scroll. Test timed out
